second commit

+++ Arac_Kiralama.Service/Concretes/BrandService.cs
@@ -5,6 +5,7 @@ using Arac_Kiralama.Repository.Contexts;
 using Arac_Kiralama.Service.Abstracts;
 using Arac_Kiralama.Models.Entity;
 using Arac_Kiralama.Service.Mappers.Brands;
+using Arac_Kiralama.Service.Exceptions.Types;
 
 
 namespace Arac_Kiralama.Service.Concretes;
@@ -22,14 +23,23 @@ public class BrandService: IBrandService
 
     public void Add(BrandAddRequestDto dto)
     {
-       Brand brand=_brandMapper.ConvertToEntity(dto);
+        bool isPresent = _brandRepository.ExistByNameAndModelYear(dto.Name,dto.ModelYear);
+        if (isPresent)
+        {
+            throw new BusinessException("Marka ve model mevcut. Tekrar eklenemez");
+        }
+        Brand brand=_brandMapper.ConvertToEntity(dto);
         _brandRepository.Add(brand);
     }
 
     public void Delete(int id)
     {
        Brand? brand=_brandRepository.GetById(id);
-        //toDo: Eğer ilgili brand bulunamazsa exeption Handling mekanizması 
+        if (brand is null)
+        {
+            throw new NotFoundException("İlgili Marka/Model bulunamadı.");
+        }
+        
         _brandRepository.Delete(brand!);
     }
 
@@ -38,6 +48,11 @@ public class BrandService: IBrandService
        List<Brand> brands =_brandRepository.GetAll();
        List<BrandResponseDto> responses = _brandMapper.ConvertToResponseList(brands);
         return responses;
+    }
+
+    public List<Brand> GetBrandsByName(string brandName)
+    {
+        return _brandRepository.GetBrandsByName(brandName);
     }
 
     public BrandResponseDto GetById(int id)
