+++ Arac_Kiralama.Service/Concretes/FuelService.cs
@@ -3,9 +3,11 @@ using Arac_Kiralama.Models.Entity;
 using Arac_Kiralama.Repository.Repositories.Abstracts;
 using Arac_Kiralama.Repository.Repositories.Concretes;
 using Arac_Kiralama.Service.Abstracts;
+using Arac_Kiralama.Service.Exceptions.Types;
 using Arac_Kiralama.Service.Mappers.Fuels;
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -25,6 +27,11 @@ namespace Arac_Kiralama.Service.Concretes
 
         public void Add(FuelAddRequestDto dto)
         {
+            bool isPresent = _fuelRepository.ExistByFuelName(dto.Name);
+            if (isPresent)
+            {
+                throw new BusinessException("Yakıt türü daha önce eklendi");
+            }
             Fuel fuel = _fuelMapper.ConvertToEntity(dto);
             _fuelRepository.Add(fuel);
         }
@@ -32,7 +39,11 @@ namespace Arac_Kiralama.Service.Concretes
         public void Delete(int id)
         {
             Fuel? fuel = _fuelRepository.GetById(id);
-            //toDo: Eğer ilgili brand bulunamazsa exeption Handling mekanizması 
+            if (fuel is null)
+            {
+                throw new NotFoundException("İlgili yakıt bulunamadı.");
+            }
+             
             _fuelRepository.Delete(fuel!);
         }
